Made errval_t a flounder builtin.
The errval type is used internally by flounder, but has since been duplicated
to be passed around through interfaces by user code.  This led to a
proliferation of ifname_errval_t types, which were used interchangably in
user-level code.  This breaks once strict typedef checking is enforced, as in
recent GCC versions.  This is a bandaid fix, and there's a larger underlying
issue still to be resolved around how to deal with globally-defined types in
flounder.

+++ include/arch/x86_64/barrelfish_kpi/registers_arch.h
@@ -38,7 +38,7 @@ registers_set_param(arch_registers_state_t *regs, uint64_t param)
 }
 
 static inline void
-registers_get_param(arch_registers_state_t *regs, uint64_t *param)
+registers_get_param(arch_registers_state_t *regs, uintptr_t *param)
 {
     registers_x86_64_get_param(regs, param);
 }
